src/test/java/com/cmpe202/individualproject/handlers/AmexTest.java
package com.cmpe202.individualproject.handlers;

import com.cmpe202.individualproject.main.CreditCardEntry;
import org.junit.Assert;
import org.junit.jupiter.api.Test;


import java.time.LocalDate;

class AmexTest {

    @Test
    void checkValidAmex() {
        Long cardNumber = 341000000000000L;
        LocalDate expDate = LocalDate.of(2030,3,20);
        String name = "eve";
        CreditCardEntry creditCardEntry = new CreditCardEntry(cardNumber,expDate,name);
        Amex amexObj = new Amex();
        String result = amexObj.checkCreditCardType(creditCardEntry);
        Assert.assertEquals(result,"American Express Card");
    }
    @Test
    void checkInvalidAmex() {
        Long cardNumber = 341000000000000L;
        LocalDate expDate = LocalDate.of(2030,3,20);
        String name = "eve";
        CreditCardEntry creditCardEntry = new CreditCardEntry(cardNumber,expDate,name);
        Amex amexObj = new Amex();
        String result = amexObj.checkCreditCardType(creditCardEntry);
        Assert.assertNotEquals(result,"Master Card");
    }

}